Add useful development amenities (#74)
* Add useful development amenities

Pinning the Python with nix sets the ground for reproducability,
Makefile allows for having self-documenting code, instead of a README
and finally direnv and the tiny script is just plain nice.

You can for example have two windows open side-by-side and copy-paste
commands between an environment that has a dt-cli install and a feature
under development. It also abstracts away poetry.