[APM-377557] Seperate buildling from building and signing
In it previous form the design proved difficult when entertaining
various scripting endeavours. Sometimes one finds oneself in need of
producing an unsinged package (that could be later signed via a
different command).

This design also allows for putting more complex logic into the build,
which will be necessary with Python extensions.