[APM-377557] Add a simple sign subcommand
The second step of seperating the build from sing. I've also taken the
liberty to support only the fused key-certificate only and postpone many
features until later. Functionally it's on par with what `build` offers,
except for dev_passphrase - but that's not interesting for this usecase.

I'm awed by how much simpler the whole things is now.