Further speed up CI
The CI had number of issues:
- the docker image was passed as file which added tons of additional
overhead
- the 'build-linux-binary' step was not using the image altogether,
effectively doing an env rebuild (the most expensive step in the entire
pipeline)

The issues were trivial to address. The only non straightforward part
was the fact that slim image doesn't have binutils, which needed to be
installed.

Passing the images between pipeline steps takes 30s, which accounts for
20% of the CI time. Therefore pararellizing the verification step won't
 speed up anything. The only improvement I deem possible
is caching pipenv packges before build and leaveraging the registry
cache (it should work now).